프로필

+++ weple-back/src/main/java/kr/co/weple/member/model/vo/Member.java
@@ -22,7 +22,7 @@ public class Member {
 	private String memberBirth;
 	private String memberGender;
 	private String memberImage;
-	private int memberLike;
+	private double memberLike;
 	private int memberGrade;
 	private String memberCategory;
 	private int memberMeet;
